# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
  x-oapi-codegen-middlewares:
    - Middleware

servers:
  - url: http://localhost
paths:
  /hello:
    get:
      summary: This is just a test endpoint to get you started. Please delete this endpoint.
      operationId: hello
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Test return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloResponse"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      summary: User Login
      operationId: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone_number:
                  type: string
                  pattern: '^\+62'
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationPhoneNumber"
                password:
                  type: string
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationPassword"
              required:
                - phone_number
                - password
      responses:
        '200':
          description: Response login API success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '400':
          description: Bad Request Code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registration:
    post:
      summary: This endpoint for registration users
      operationId: Registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone_number:
                  type: string
                  minLength: 10
                  maxLength: 20
                  example: "+628577455545"
                  pattern: '^\+62'
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationPhoneNumber"
                full_name:
                  type: string
                  minLength: 3
                  maxLength: 60
                  example: "sawit Pro"
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationFullName"
                password:
                  type: string
                  minLength: 6
                  maxLength: 64
                  example: "asdqwe1A@"
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationPassword"
              required:
                - phone_number
                - full_name
                - password
      responses:
        '200':
          description: Response Registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /get-profile:
    get:
      summary: This is just a get my profile endpoint.
      operationId: GetMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Test return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMyProfileResponse"
        '400':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Forbiden Code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /update-profile:
    patch:
      summary: This is just for update profile data
      operationId: UpdateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone_number:
                  type: string
                  minLength: 10
                  maxLength: 20
                  example: "+628577455545"
                  pattern: '^\+62'
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationPhoneNumber"
                full_name:
                  type: string
                  minLength: 3
                  maxLength: 60
                  example: "sawit Pro"
                  x-oapi-codegen-extra-tags:
                    validate: "required,validationFullName"
              required:
                - phone_number
                - full_name
      responses:
        '200':
          description: Response Update Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseUpdateProfile"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: Forbiden Code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HelloResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    LoginResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            user_id:
              type: integer
            auth_jwt:
              type: string
      example:
        code: 200
        message: success
        data:
          user_id: 1
          auth_jwt: "auth rsa 256"
    RegistrationResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            user_id:
              type: integer
              description: user id registered
      example:
        code: 200
        message: success
        data:
          user_id: 1
    ResponseUpdateProfile:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            user_id:
              type: integer
              description: user id affected by updated
      example:
        code: 200
        message: success
        data:
          user_id: 1
    GetMyProfileResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            full_name:
              type: string
              description: full_name is data catch from database by token
            phone_number:
              type: string
              description: phone_number is data catch from database by token
      example:
        code: 200
        message: success
        data:
          full_name: "sawit pro"
          phone_number: "+6237087630662"
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        errors:
          type: string
          properties:
            list_error:
              type: array
              items:
                type: string
                description: List of errors
      example:
        code: 400
        errors:
          list_error:
            - Password does not meet the password requirements
        message: invalid request
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using RS256